test cases and captcha code login

+++ src/main/java/com/hellodoctor/servicesImpl/HelloDoctorServicesImplements.java
@@ -5,17 +5,20 @@ import java.util.Base64;
 import java.util.Date;
 import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.Optional;
 
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.util.ObjectUtils;
+
 import com.hellodoctor.constant.Constant;
 import com.hellodoctor.entities.Appointment;
 import com.hellodoctor.entities.Doctor;
 import com.hellodoctor.entities.HospitalAddress;
 import com.hellodoctor.entities.HospitalsDetails;
 import com.hellodoctor.entities.Users;
+import com.hellodoctor.exception.AlreadyUseException;
 import com.hellodoctor.exception.BusinessException;
 import com.hellodoctor.exception.EmptyInputException;
 import com.hellodoctor.exception.RecordNotFoundException;
@@ -28,6 +31,7 @@ import com.hellodoctor.requestdto.DoctorUpdateDto;
 import com.hellodoctor.requestdto.RequestDto;
 import com.hellodoctor.responsedto.AppointmentResponceDto;
 import com.hellodoctor.services.HelloDoctorServices;
+
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
@@ -49,27 +53,32 @@ public class HelloDoctorServicesImplements implements HelloDoctorServices {
 	@Override
 	public Doctor updateDoctor(DoctorUpdateDto doctorUpdateDto, String doctorEmail) {
 		log.info("inside Doctor Update Service");
-		Doctor findBydoctorEmail = doctorRepository.findBydoctorEmail(doctorEmail);
-		Users findById = usersRepository.findById(findBydoctorEmail.getUserId().getUserId()).orElse(null);
-		// Doctor doctor = new Doctor();
-		if (findBydoctorEmail != null) {
-			findBydoctorEmail.setDoctorName(doctorUpdateDto.getDoctorName());
-			findBydoctorEmail.setDoctorMobileNumber(doctorUpdateDto.getDoctorMobileNumber());
-			findBydoctorEmail.setDoctorPassword(doctorUpdateDto.getDoctorPassword());
-			findBydoctorEmail.setHospitalName(doctorUpdateDto.getHospitalName());
-			findBydoctorEmail.setUpdateDate(new Date());
-			log.info("saving Doctor Info In Doctor Entity");
-			doctorRepository.save(findBydoctorEmail);
+		Doctor doctorDetails = doctorRepository.findBydoctorEmail(doctorEmail);
+		if(ObjectUtils.isEmpty(doctorDetails)) {
+			throw new RecordNotFoundException(Constant.DOCTOR_NOT_FOUND +doctorEmail);
 		}
-		if (findById != null) {
-			findById.setEmail(findBydoctorEmail.getDoctorEmail());
-			findById.setMobile(doctorUpdateDto.getDoctorMobileNumber());
-			findById.setPassword(doctorUpdateDto.getDoctorPassword());
-			usersRepository.save(findById);
+		Users user = usersRepository.findById(doctorDetails.getUserId().getUserId()).orElse(null);
+		// Doctor doctor = new Doctor();
+		if (doctorDetails != null) {
+			doctorDetails.setDoctorName(doctorUpdateDto.getDoctorName());
+			doctorDetails.setDoctorMobileNumber(doctorUpdateDto.getDoctorMobileNumber());
+			doctorDetails.setDoctorPassword(Base64.getEncoder().encodeToString(doctorUpdateDto.getDoctorPassword()
+					.toString().getBytes()));
+			doctorDetails.setHospitalName(doctorUpdateDto.getHospitalName());
+			doctorDetails.setUpdateDate(new Date());
+			log.info("saving Doctor Info In Doctor Entity");
+			doctorRepository.save(doctorDetails);
+		}
+		if (user != null) {
+			user.setEmail(doctorDetails.getDoctorEmail());
+			user.setMobile(doctorUpdateDto.getDoctorMobileNumber());
+			user.setPassword(Base64.getEncoder().encodeToString(doctorUpdateDto.getDoctorPassword()
+					.toString().getBytes()));
+			usersRepository.save(user);
 		} else {
 			throw new EmptyInputException(Constant.EXCEPTIONEMAILVALIDATION);
 		}
-		return findBydoctorEmail;
+		return doctorDetails;
 
 	}
 	// doctorUpdateServiceEnd
@@ -133,8 +142,15 @@ public class HelloDoctorServicesImplements implements HelloDoctorServices {
 //			log.info("inside adddoctor hosptialServicesImplements ExceptionBlock");
 //			throw new BusinessException(ex.getMessage());
 //		}
+		Doctor doctorEmail = doctorRepository.findBydoctorEmail(requestDto.getDoctorEmail());
+		if(!ObjectUtils.isEmpty(doctorEmail)) {
+			throw new AlreadyUseException(Constant.EMAIL_USE);
+		}
 		
-		
+		Doctor mobileNumber = doctorRepository.findByDoctorMobileNumber(requestDto.getDoctorMobileNumber());
+		if(!ObjectUtils.isEmpty(mobileNumber)) {
+			throw new AlreadyUseException(Constant.MOBILE_NUMBER_USE);
+		}
 		Doctor doctor = new Doctor();
 		
 		BeanUtils.copyProperties(requestDto, doctor);
@@ -142,13 +158,15 @@ public class HelloDoctorServicesImplements implements HelloDoctorServices {
 				(Base64.getEncoder().encodeToString(requestDto.getDoctorPassword().toString().getBytes())));
 		doctor.setRegisterDate(new Date());
 		doctor.setRole(Constant.DOCTORROLE);
+		log.info("fetching hospital Details from dataBase");
 		HospitalsDetails hospitalsDetails = hospitalsDetailsRepository.findByhospitalName(requestDto.getHospitalName());
 		
 		if(ObjectUtils.isEmpty(hospitalsDetails)) {
 			log.error("current address not found  " + requestDto.getHospitalName());
-			throw new RecordNotFoundException("current address is  not registerd");
+			throw new RecordNotFoundException(Constant.ADDRESS_NOT_FOUND);
 		}
 		doctor.setHospitalsDetails(hospitalsDetails);
+		log.info("Doctor save in dataBase");
 		Doctor savedoctor = doctorRepository.save(doctor);
 		
 		log.info("calling user object");
@@ -162,9 +180,9 @@ public class HelloDoctorServicesImplements implements HelloDoctorServices {
 																											// //
 																											// password
 		users.setRole(Constant.DOCTORROLE);
-		log.info("saving doctor");
-		log.info("saving user");
+		log.info("user save in user table");
 		usersRepository.save(users);
+		log.info("add Doctor method return");
 		return doctor;
 	}
 
@@ -176,7 +194,12 @@ public class HelloDoctorServicesImplements implements HelloDoctorServices {
 			log.error("addhopsital Error called");
 			throw new EmptyInputException(Constant.EXCEPTION601);
 		}
+		
+		HospitalsDetails hospitalName = hospitalsDetailsRepository.findByhospitalName(requestDto.getHospitalName());
 
+		if(!ObjectUtils.isEmpty(hospitalName)) {
+			throw new AlreadyUseException(Constant.HOSPITAL_NAME_USE);
+		}
 		log.info("inside try method of addhospital service");
 		HospitalsDetails hospitalsDetails = new HospitalsDetails();
 		hospitalsDetails.setHospitalName(requestDto.getHospitalName());
@@ -206,31 +229,33 @@ public class HelloDoctorServicesImplements implements HelloDoctorServices {
 
 	@Override
 	public HospitalsDetails getHospitalById(Long hospitalId) {
-		try {
-			return hospitalsDetailsRepository.findById(hospitalId).get();
-		} catch (IllegalArgumentException e) {
-			log.info("insideIllegalArgumentExceptionCatchBlock");
-			throw new BusinessException(Constant.EXCEPTION606 + e.getMessage());
-		} catch (NoSuchElementException e) {
-			throw new BusinessException(Constant.EXCEPTION607 + e.getMessage());
-		} catch (Exception e) {
-			throw new BusinessException(Constant.EXCEPTION608 + e.getMessage());
+		
+		HospitalsDetails hospitalsDetails = hospitalsDetailsRepository.findById(hospitalId).orElse(null);
+		
+		if(ObjectUtils.isEmpty(hospitalsDetails)) {
+			throw new RecordNotFoundException(Constant.HOSPITAL_NOT_FOUND);
 		}
+		return hospitalsDetails;
+//		try {
+//			return hospitalsDetailsRepository.findById(hospitalId).get();
+//		} catch (IllegalArgumentException e) {
+//			log.info("insideIllegalArgumentExceptionCatchBlock");
+//			throw new BusinessException(Constant.EXCEPTION606 + e.getMessage());
+//		} catch (NoSuchElementException e) {
+//			throw new BusinessException(Constant.EXCEPTION607 + e.getMessage());
+//		} catch (Exception e) {
+//			throw new BusinessException(Constant.EXCEPTION608 + e.getMessage());
+//		}
 	}
 
 	// get doctor by email
 	@Override
 	public Doctor findBydoctorEmail(String doctorEmail) {
-		try {
-			return doctorRepository.findBydoctorEmail(doctorEmail);
-		} catch (IllegalArgumentException e) {
-			log.info("insideIllegalArgumentExceptionCatchBlock");
-			throw new BusinessException(Constant.EXCEPTION621 + e.getMessage());
-		} catch (NoSuchElementException e) {
-			throw new BusinessException(Constant.EXCEPTION620 + e.getMessage());
-		} catch (Exception e) {
-			throw new BusinessException(Constant.EXCEPTION608 + e.getMessage());
+		Doctor doctor = doctorRepository.findBydoctorEmail(doctorEmail);
+		if(ObjectUtils.isEmpty(doctor)) {
+			throw new RecordNotFoundException(Constant.DOCTOR_NOT_FOUND+doctorEmail);
 		}
+			return doctor;
 	}
 
 	// get doctor by email end
@@ -254,16 +279,13 @@ public class HelloDoctorServicesImplements implements HelloDoctorServices {
 	// get by addressId
 	@Override
 	public HospitalAddress getHospitalAddressById(Long addressId) {
-		try {
-			return hospitalAddressRepository.findById(addressId).get();
-		} catch (IllegalArgumentException e) {
-			log.info("insideIllegalArgumentExceptionCatchBlock");
-			throw new BusinessException(Constant.EXCEPTION606 + e.getMessage());
-		} catch (NoSuchElementException e) {
-			throw new BusinessException(Constant.EXCEPTION607 + e.getMessage());
-		} catch (Exception e) {
-			throw new BusinessException(Constant.EXCEPTION608 + e.getMessage());
+
+		HospitalAddress address = hospitalAddressRepository.findById(addressId).orElse(null);
+		if (ObjectUtils.isEmpty(address)) {
+			throw new RecordNotFoundException(Constant.HOSPITAL_ADDRESSNOT_NOT_FOUND);
 		}
+
+		return address;
 	}
 
 	// get by addressId end
@@ -303,34 +325,23 @@ public class HelloDoctorServicesImplements implements HelloDoctorServices {
 	// delete by DoctorById
 	@Override
 	public void deletedoctorById(Long doctorId) {
-		try {
-			doctorRepository.deleteById(doctorId);
-
-		} catch (IllegalArgumentException e) {
-			log.info("inside addEmployee AttendenceServicesImplements IllegalArgumentExceptionCatchBlock");
-			throw new BusinessException(Constant.EXCEPTION609 + e.getMessage());
-		} catch (Exception e) {
-			throw new BusinessException(Constant.EXCEPTION613 + e.getMessage());
+		Optional<Doctor> optional = doctorRepository.findById(doctorId);
+		
+		if(ObjectUtils.isEmpty(optional)) {
+			throw new RecordNotFoundException("Record not found " +doctorId);
 		}
-
+			doctorRepository.deleteById(doctorId);
 	}
 	// delete by DoctorById End
 
 	// get hospital by name
 	@Override
-	public List<HospitalsDetails> findByhospitalName(String hospitalName) {
-//		List<HospitalsDetails> hospitallist = null;
-//		hospitallist = hospitalsDetailsRepository.findByhospitalName(hospitalName);
-//		try {
-//			hospitallist = hospitalsDetailsRepository.findByhospitalName(hospitalName);
-//		} catch (Exception e) {
-//			throw new BusinessException(Constant.EXCEPTION618);
-//		}
-//		if (hospitallist.isEmpty())
-//			throw new BusinessException(Constant.EXCEPTION604);
-//		log.info("doctor list object");
-//		return hospitallist;
-		return null;
+	public HospitalsDetails findByhospitalName(String hospitalName) {
+		HospitalsDetails hospitalsDetails = hospitalsDetailsRepository.findByhospitalName(hospitalName);
+		if(ObjectUtils.isEmpty(hospitalsDetails)) {
+			throw new RecordNotFoundException("hospital name not available");
+		}
+		return hospitalsDetails;
 	}
 
 	@Override
@@ -372,14 +383,17 @@ public class HelloDoctorServicesImplements implements HelloDoctorServices {
 	@Override
 	public List<AppointmentResponceDto> getAppointment(String email) {
 		Doctor doctor = doctorRepository.findBydoctorEmail(email);
-List<AppointmentResponceDto> appointmentResponceDto = new ArrayList<>();
-		
+		if (ObjectUtils.isEmpty(doctor)) {
+			throw new RecordNotFoundException(Constant.DOCTOR_NOT_FOUND + email);
+		}
+		List<AppointmentResponceDto> appointmentResponceDto = new ArrayList<>();
+
 		log.info(" fetching appointment ");
 		List<Appointment> appointmentDetails = appointmentRepository.findByDoctorEmail(doctor.getDoctorEmail());
-		if(ObjectUtils.isEmpty(appointmentDetails)) {
-			throw new RecordNotFoundException("appointment is empty");
+		if (ObjectUtils.isEmpty(appointmentDetails)) {
+			throw new RecordNotFoundException(Constant.APPOINTMENT_NOT_FOUND);
 		}
-		for(Appointment appDeatiels :appointmentDetails) {
+		for (Appointment appDeatiels : appointmentDetails) {
 			AppointmentResponceDto dto = new AppointmentResponceDto();
 			dto.setDoctorId(appDeatiels.getDoctor().getDoctorId());
 			dto.setPatientId(appDeatiels.getPatient().getPatientId());
